+++ feature/utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/sdcard/SDCardActivity.kt
@@ -32,19 +32,15 @@ class SDCardActivity : CommonActivity() {
     }
 
     override fun bindItems(): MutableList<CommonItem<*>> {
-        val mountedSDCardPath = SDCardUtils.getMountedSDCardPath()
-        val sizeItems = CollectionUtils.collect(mountedSDCardPath) { input ->
-            val totalInfo = "total: " + ConvertUtils.byte2FitMemorySize(SDCardUtils.getTotalSize(input))
-            val availableInfo = "available: " + ConvertUtils.byte2FitMemorySize(SDCardUtils.getAvailableSize(input))
-            CommonItemTitle(input, "$totalInfo, $availableInfo")
-        }
-        val result: ArrayList<CommonItem<*>> = CollectionUtils.newArrayList(
+        return CollectionUtils.newArrayList(
                 CommonItemTitle("isSDCardEnableByEnvironment", SDCardUtils.isSDCardEnableByEnvironment().toString()),
                 CommonItemTitle("getSDCardPathByEnvironment", SDCardUtils.getSDCardPathByEnvironment()),
                 CommonItemTitle("getSDCardInfo", SDCardUtils.getSDCardInfo().toString()),
-                CommonItemTitle("getMountedSDCardPath", mountedSDCardPath.toString())
+                CommonItemTitle("getMountedSDCardPath", SDCardUtils.getMountedSDCardPath().toString()),
+                CommonItemTitle("getExternalTotalSize", ConvertUtils.byte2FitMemorySize(SDCardUtils.getExternalTotalSize(), 2)),
+                CommonItemTitle("getExternalAvailableSize", ConvertUtils.byte2FitMemorySize(SDCardUtils.getExternalAvailableSize(), 2)),
+                CommonItemTitle("getInternalTotalSize", ConvertUtils.byte2FitMemorySize(SDCardUtils.getInternalTotalSize(), 2)),
+                CommonItemTitle("getInternalAvailableSize", ConvertUtils.byte2FitMemorySize(SDCardUtils.getInternalAvailableSize(), 2))
         )
-        result.addAll(sizeItems)
-        return result
     }
 }